updated_commit

+++ shop/products/forms.py
@@ -1,9 +1,9 @@
 from flask_wtf.file import FileAllowed, FileField, FileRequired
-from wtforms import Form, IntegerField, StringField, BooleanField, TextAreaField, validators
+from wtforms import Form, IntegerField, StringField, DecimalField, TextAreaField, validators
 
 class Addproducts(Form):
     name = StringField('Name', [validators.DataRequired()])
-    price = IntegerField('Price',[validators.DataRequired()])
+    price = DecimalField('Price',[validators.DataRequired()])
     discount = IntegerField('Discount', default=0)
     stock = IntegerField('Stock', [validators.DataRequired()])
     desc =TextAreaField('Description', [validators.DataRequired()])
